stats: fix dubious cast

+++ modules/misc/stats.c
@@ -152,7 +152,8 @@ static int Demux( demux_t *p_demux )
     p_block->i_dts = p_block->i_pts =
         date_Increment( &p_sys->pts, kBufferSize );
 
-    msg_Dbg( p_demux, "demux got %d ms offset", (int)(mdate() - *(mtime_t *)p_block->p_buffer) / 1000 );
+    msg_Dbg( p_demux, "demux got %"PRId64" ms offset",
+             (mdate() - *(mtime_t *)p_block->p_buffer) / 1000 );
 
     //es_out_SetPCR( p_demux->out, p_block->i_pts );
